Changes to start using poker_end
Players hands were not being cleared because poker_end wasn't being called.
Fixed not being able to change a class variable for appending the player's hand to the end of the deck in the end of the match
  Now doing the same operation from a procedure inside of the class
Changed some returns and other procedures to nudge them to comply with what we want poker_end to do.
I'm not happy with how the flow of poker_end, poker_init sort of work. It's not elegant between starting a game (multiple hands with changing player's cash) and starting different rounds.
  I will fix soon to make more sense. Or perhaps we should consider if we want the RL to even care about money.